[-] Factorize the parsing params for primaryCompositeKey
