+++ services/resource-remover.js
@@ -1,21 +1,10 @@
 'use strict';
-var Interface = require('forest-express');
-var CompositeKeysManager = require('./composite-keys-manager');
+var ResourceFinder = require('./resource-finder');
 
 function ResourceRemover(model, params) {
-  var schema = Interface.Schemas.schemas[model.name];
-
   this.perform = function () {
-    var where = {};
-    var compositeKeysManager = new CompositeKeysManager(model, schema, params);
-
-    if (schema.isCompositePrimary) {
-      where = compositeKeysManager.getRecordConditions(params.recordId);
-    } else {
-      where[schema.idField] = params.recordId;
-    }
-    return model
-      .find({ where: where })
+    return new ResourceFinder(model, params)
+      .perform()
       .then(function (record) {
         return record.destroy();
       });
